Rename getEntrypointFilesGroupedByExtension
In the following commit I'd like to enable the inclusion of
`namedChunkGroups` in the outputted manifest file. It will work exactly
the same as `entrypoints` but include all named chunk groups - including
dynamically loaded ones.

This commit renames the `getEntrypointFilesGroupedByExtension` function
to `getChunkGroupFilesGroupedByExtension` so it makes sense for both
`compilation.entrypoints` and `compilation.namedChunkGroups`.