Optionally include namedChunkGroups
This commit enables the inclusion of `namedChunkGroups` in the outputted
manifest file. It works similarly to `entrypoints`, but includes all
named chunk groups – including dynamically loaded ones.

For example, if we have an entrypoint like this:

```
import(/* webpackChunkName: "some-dynamic-import" */ 'some-module');
```

…using the `namedChunkGroups` option will produce a manifest like:

```
{
  "namedChunkGroups": {
    "some-entrypoint": {
      "js": [
        "some-entrypoint.js"
      ]
    },
    "some-dynamic-import": {
      "js": [
        "some-dependency.js",
        "some-module.js"
      ],
      "css": [
        "some-dependency.css",
        "some-module.css"
      ]
    }
  }
}
```

This could be useful for adding stylesheet link tags or preloading
assets for modules we know are going to be dynamically loaded.